helpers/token_helpers.py
from datetime import datetime, timedelta
from typing import Optional

from jose import jwt

from models.account_model import User

SECRET_KEY = "ffbd262d207c97f5e80502c183fa2efbc5109494031d80b890c44748d1be2391"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


