requirements.txt
SQLAlchemy>=1.4.35
uvicorn>=0.17.6
fastapi>=0.75.2
pydantic>=1.9.0
pika>=1.2.0
passlib>=1.7.4
psycopg2>=2.9.2
jwt>=1.3.1
python-jose>=3.3.0
python-multipart>=0.0.5
bcrypt>=3.2.2
